Separate rpmParse functions into different headers to fix CentOS compilation
